Move RequestDirector
